Move tournament files in dedicated folder.
