+++ src/components/Tournament/Admin/TournamentAdminSwiss.vue
@@ -7,6 +7,9 @@ import {
   required,
 } from '@vuelidate/validators';
 import { computed, reactive, ref } from 'vue';
+import FormField from '@/components/FormField.vue';
+import Modal from '@/components/Modal.vue';
+import AdminMatch from '@/components/Tournament/Admin/AdminMatch.vue';
 import {
   BestofType,
   MatchTypeEnum,
@@ -14,10 +17,6 @@ import {
 import type { TournamentDeref } from '@/models/tournament';
 import { useNotificationStore } from '@/stores/notification.store';
 import { groupBy, useTournamentStore } from '@/stores/tournament.store';
-
-import AdminMatch from './AdminMatch.vue';
-import FormField from './FormField.vue';
-import Modal from './Modal.vue';
 
 const { tournament } = defineProps<{
   tournament: TournamentDeref;
