+++ src/components/Tournament/GroupTable.vue
