+++ src/components/Tournament/SeatingPlan.vue
