+++ src/components/Tournament/TournamentCard.vue
