+++ src/components/Tournament/TournamentInfo.vue
