+++ src/components/Tournament/TournamentRules.vue
