+++ src/views/Tournament/TournamentDetail.vue
