+++ src/views/Tournament/TournamentRegister.vue
